Cell column index doesn't necessarily match up with how cells are read into list
When reading a row from a spreadsheet if a column is left blank there is a possibility to read fewer columns than you would see in excel.  The blank cell is simply omitted from the row.  To remedy it's important to make sure column index in the mapped dictionary actually matches up with what column the cell is really from.